bot/main.py
import os
from fastapi import FastAPI, Request
from telegram.ext import Application
from bot.telegram_router import setup_handlers
from contextlib import asynccontextmanager
from bot.error_middleware import add_error_middleware


@asynccontextmanager
async def lifespan(app: FastAPI):
    """Lifespan context manager for the FastAPI app"""
    # Initialize the bot on startup
    app.state.telegram_bot = await initialize_bot()
    yield
    # Shutdown the bot on app shutdown
    if app.state.telegram_bot:
        await app.state.telegram_bot.shutdown()

# Initialize Telegram Bot application


async def initialize_bot():
    """Initialize the Telegram bot application"""
    from config import Config

    try:
        telegram_bot = Application.builder().token(
            Config.TELEGRAM_BOT_TOKEN).build()
        setup_handlers(telegram_bot)
        await telegram_bot.initialize()
        return telegram_bot
    except Exception:
        # In test environments, we might not have a real token, so return a stub
        if os.environ.get("TESTING") == "True":
            return None
        raise


def build_app():
    """Build and configure the FastAPI application"""
    app = FastAPI(lifespan=lifespan)

    # Add error handling middleware
    add_error_middleware(app)

    @app.get("/")
    async def root():
        """Root endpoint to indicate bot is active."""
        return {
            "status": "ok",
            "message": ("Bot is active. Webhook is configured at /webhook. "
                        "Health check at /health.")
        }

    @app.post("/webhook")
    async def webhook(request: Request):
        """Handle Telegram webhook updates"""
        # Get update data
        update_data = await request.json()

        # Handle the update differently in test mode
        if os.environ.get("TESTING") == "True" or app.state.telegram_bot is None:
            # For testing, just return OK
            # The test fixtures will handle mocking the appropriate components
            return {"status": "ok"}

        # Use the actual bot to process the update
        await app.state.telegram_bot.update_queue.put(update_data)
        return {"status": "ok"}

    @app.get("/health")
    async def health():
        """Health check endpoint for Cloud Run"""
        return {"status": "ok"}

    return app


# Create the app instance
app = build_app()

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(
        "bot.main:app",
        host="0.0.0.0",
        port=int(os.getenv("PORT", 8080))
    )
