requirements.txt
python-telegram-bot==20.7
google-cloud-firestore==2.13.1
openai>=1.79.0
google-generativeai==0.3.1
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
docker==7.0.0
requests==2.31.0 